update `PySet_GET_SIZE` and similar for free-threaded Python (#6230)
* update `PySet_GET_SIZE` and similar for free-threaded Python

* split `cpython` portion of `setobject.h` off
