+++ newsfragments/6230.fixed.md
@@ -1,0 +1,1 @@
+Fix FFI definitions `PyByteArray_GET_SIZE`, `PyList_GET_SIZE`, and `PySet_GET_SIZE` to use an atomic load for free-threaded Python.
